Properly position the cursor at the first stop in some cases when inserting snippets without editor support (#2017)
Applies when:

- there is text on the line after the inserted snippet
- the first stop in the snippet is on the last (or only) line

Also documents an assumption in `move_to_correct_column` that caused the
incorrect behavior.

Per discussion on Slack, we may want to change the behavior of
`move_to_correct_column` to not go to the end of the line, but this
involves different tradeoffs that should be discussed further.